refactor nested route registration
